+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# sin publicar
+- Changed MLBusinessItemDescriptionView title number of lines to 3
+
 # v1.42.0
 🚀 1.42.0 🚀
 - Fix in how FlexCoverCarousel Cells were reusing data.