+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# sin publicar
+# v1.43.0
+🚀 1.43.0 🚀
 - Changed MLBusinessItemDescriptionView title number of lines to 3
 - Fix FlexCoverCarousel cards when receiving nil in text parameters.
 - Fix for cases when we don't have the labels' color values from be